Oppdater eksisterende nasjonale triggere
