+++ src/schemas/ba-sak/begrunnelse/nasjonaleTriggere/borMedSøkerTriggere.ts
@@ -1,5 +1,6 @@
 import { BegrunnelseDokumentNavn, SanityTyper } from '../../../../util/typer';
 import { borMedSøkerTriggerTyper, Vilkår, vilkårTriggerTilMenynavn } from '../typer';
+import { hentNasjonaleTriggereRegler } from './utils';
 
 export const borMedSøkerTriggere = {
   title: 'Triggere for "Bor med søker"',
@@ -10,4 +11,5 @@ export const borMedSøkerTriggere = {
     list: borMedSøkerTriggerTyper.map(trigger => vilkårTriggerTilMenynavn[trigger]),
   },
   hidden: ({ document }) => !(document.vilkaar && document.vilkaar.includes(Vilkår.BOR_MED_SOKER)),
+  validation: rule => hentNasjonaleTriggereRegler(rule),
 };
