+++ src/schemas/ba-sak/begrunnelse/nasjonaleTriggere/bosattIRiketTriggere.ts
@@ -1,5 +1,6 @@
 import { BegrunnelseDokumentNavn, SanityTyper } from '../../../../util/typer';
 import { bosattIRiketTriggerTyper, Vilkår, vilkårTriggerTilMenynavn } from '../typer';
+import { hentNasjonaleTriggereRegler } from './utils';
 
 export const bosattIRiketTriggere = {
   title: 'Triggere for "Bosatt i riket"',
@@ -10,4 +11,5 @@ export const bosattIRiketTriggere = {
     list: bosattIRiketTriggerTyper.map(trigger => vilkårTriggerTilMenynavn[trigger]),
   },
   hidden: ({ document }) => !(document.vilkaar && document.vilkaar.includes(Vilkår.BOSATT_I_RIKET)),
+  validation: rule => hentNasjonaleTriggereRegler(rule),
 };
