+++ src/schemas/ba-sak/begrunnelse/nasjonaleTriggere/endretUtbetalingPeriodeTrigger.ts
@@ -1,6 +1,8 @@
 import { BegrunnelseDokumentNavn, SanityTyper } from '../../../../util/typer';
 import { endretUtbetalingsperioderTriggereValg } from '../typer';
 import { erEndretUtbetaling } from './endringsårsakTrigger';
+import { erEøsBegrunnelse } from '../EØSTriggere/utlis';
+import { hentNasjonaleTriggereRegler } from './utils';
 
 export const endretUtbetalingsperiodeTriggere = {
   title: 'Endret utbetalingsperiode triggere',
@@ -10,5 +12,6 @@ export const endretUtbetalingsperiodeTriggere = {
   options: {
     list: endretUtbetalingsperioderTriggereValg,
   },
-  hidden: ({ document }) => !erEndretUtbetaling(document),
+  hidden: ({ document }) => !erEndretUtbetaling(document) || erEøsBegrunnelse(document),
+  validation: rule => hentNasjonaleTriggereRegler(rule),
 };
