+++ src/schemas/ba-sak/begrunnelse/nasjonaleTriggere/giftPartnerskapTriggere.ts
@@ -1,5 +1,6 @@
 import { giftPartnerskapTriggerTyper, Vilkår, vilkårTriggerTilMenynavn } from '../typer';
 import { BegrunnelseDokumentNavn, SanityTyper } from '../../../../util/typer';
+import { hentNasjonaleTriggereRegler } from './utils';
 
 export const giftPartnerskapTriggere = {
   title: 'Triggere for "Gift partnerskap"',
@@ -11,4 +12,5 @@ export const giftPartnerskapTriggere = {
   },
   hidden: ({ document }) =>
     !(document.vilkaar && document.vilkaar.includes(Vilkår.GIFT_PARTNERSKAP)),
+  validation: rule => hentNasjonaleTriggereRegler(rule),
 };
