+++ src/schemas/ba-sak/begrunnelse/nasjonaleTriggere/lovligOppholdTriggere.ts
@@ -1,5 +1,6 @@
 import { BegrunnelseDokumentNavn, SanityTyper } from '../../../../util/typer';
 import { lovligOppholdTriggerTyper, Vilkår, vilkårTriggerTilMenynavn } from '../typer';
+import { hentNasjonaleTriggereRegler } from './utils';
 
 export const lovligOppholdTriggere = {
   title: 'Triggere for "Lovlig opphold"',
@@ -10,4 +11,5 @@ export const lovligOppholdTriggere = {
     list: lovligOppholdTriggerTyper.map(trigger => vilkårTriggerTilMenynavn[trigger]),
   },
   hidden: ({ document }) => !(document.vilkaar && document.vilkaar.includes(Vilkår.LOVLIG_OPPHOLD)),
+  validation: rule => hentNasjonaleTriggereRegler(rule),
 };
