+++ src/schemas/ba-sak/begrunnelse/nasjonaleTriggere/utvidetBarnetrygdTriggere.ts
@@ -1,5 +1,7 @@
 import { BegrunnelseDokumentNavn, SanityTyper } from '../../../../util/typer';
 import { utvidetBarnetrygdTriggertyper, Vilkår, vilkårTriggerTilMenynavn } from '../typer';
+import { erEøsBegrunnelse } from '../EØSTriggere/utlis';
+import { hentNasjonaleTriggereRegler } from './utils';
 
 export const utvidetBarnetrygdTriggere = {
   title: 'Utvidet barnetrygd triggere',
@@ -10,5 +12,7 @@ export const utvidetBarnetrygdTriggere = {
     list: utvidetBarnetrygdTriggertyper.map(trigger => vilkårTriggerTilMenynavn[trigger]),
   },
   hidden: ({ document }) =>
-    !(document.vilkaar && document.vilkaar.includes(Vilkår.UTVIDET_BARNETRYGD)),
+    !(document.vilkaar && document.vilkaar.includes(Vilkår.UTVIDET_BARNETRYGD)) ||
+    erEøsBegrunnelse(document),
+  validation: rule => hentNasjonaleTriggereRegler(rule),
 };
