+++ src/schemas/ba-sak/begrunnelse/nasjonaleTriggere/øvrigeTriggere.ts
@@ -1,6 +1,8 @@
 import { BegrunnelseDokumentNavn, SanityTyper } from '../../../../util/typer';
 import { erEndretUtbetaling } from './endringsårsakTrigger';
 import { vilkårTriggerTilMenynavn, øvrigeTriggertyper } from '../typer';
+import { erEøsBegrunnelse } from '../EØSTriggere/utlis';
+import { hentNasjonaleTriggereRegler } from './utils';
 
 export const øvrigeTriggere = {
   title: 'Øvrige triggere',
@@ -10,5 +12,6 @@ export const øvrigeTriggere = {
   options: {
     list: øvrigeTriggertyper.map(trigger => vilkårTriggerTilMenynavn[trigger]),
   },
-  hidden: ({ document }) => erEndretUtbetaling(document),
+  hidden: ({ document }) => erEndretUtbetaling(document) || erEøsBegrunnelse(document),
+  validation: rule => hentNasjonaleTriggereRegler(rule),
 };
